Content changes and our Market resonsive added
